Module Core

Contains everything regarding services, their lifecycles and request contexts. You want to touch this only if you develop Sihl services yourself.

module Container = Core__.Core_container

module Ctx = Core__.Core_ctx