Module Email

Use this module to send single emails either immediately or later sending a lot of emails at once. Email template support is also provided.

exception Exception of string

module Template : sig ... end

module DevInbox = Email__.Model.DevInbox

type t = Email__.Model.t

val make : sender:string -> recipient:string -> subject:string -> text_content:string -> html_content:string -> ?⁠cc:string list -> ?⁠bcc:string list -> html:bool -> ?⁠template_id:string -> ?⁠template_data:Template.Data.t -> unit -> t

val template_data : t -> Template.Data.t

val template_id : t -> string option

val html : t -> bool

val bcc : t -> string list

val cc : t -> string list

val text_content : t -> string

val html_content : t -> string

val subject : t -> string

val recipient : t -> string

val sender : t -> string

val set_text_content : string -> t -> t

val set_html_content : string -> t -> t

val pp : Stdlib.Format.formatter -> t -> unit

val show : t -> string

val equal : t -> t -> bool

val to_yojson : t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> t Ppx_deriving_yojson_runtime.error_or

module Sig : sig ... end