Module Email__.Email_service

module Sig = Email__.Email_service_sig

module EnvConfigProvider : functor (Config : Configuration.Service.Sig.SERVICE) -> Sig.CONFIG_PROVIDER_SMTP

module Template : sig ... end

module Make : sig ... end

module MakeDelayed : functor (Log : Log.Service.Sig.SERVICE) -> functor (EmailService : Sig.SERVICE) -> functor (DbService : Data.Db.Service.Sig.SERVICE) -> functor (QueueService : Queue.Service.Sig.SERVICE) -> Sig.SERVICE

Use this functor to create an email service that sends emails using the job queue. This is useful if you need to answer a request quickly while sending the email in the background