Module Email__Email_service.MakeDelayed

Use this functor to create an email service that sends emails using the job queue. This is useful if you need to answer a request quickly while sending the email in the background

Parameters

Log : Log.Service.Sig.SERVICE

EmailService : Sig.SERVICE

DbService : Data.Db.Service.Sig.SERVICE

QueueService : Queue.Service.Sig.SERVICE

Signature

include Core.Container.SERVICE

val lifecycle : Core.Container.Lifecycle.t

module Template : Sig.TEMPLATE_SERVICE

A template service to manage email templates.

val send : Core.Ctx.t -> Email__.Email_core.t -> unit Lwt.t

Send email.

val bulk_send : Core.Ctx.t -> Email__.Email_core.t list -> unit Lwt.t

Send multiple emails. If sending of one of them fails, the function fails.