Module Middleware

A middleware is a function that takes a Http.Route.handler and returns a Http.Route.handler. It is typically used to add content to the request context that is valid only during a request.

type t = Middleware__.Middleware_core.t

type stack = Middleware__.Middleware_core.stack

A middleware stack is a list of middlewares.

val apply : t -> Http.Route.handler -> Http.Route.handler

apply m h applies the middleware m with the handler h, so that the middleware logic wraps the handler.

val apply_stack : stack -> Http.Route.t -> Http.Route.t

module Clickjacking = Middleware__.Middleware_clickjacking

module Cookie = Middleware__.Middleware_cookie

module Csrf = Middleware__.Middleware_csrf

module Db = Middleware__.Middleware_db

module Error = Middleware__.Middleware_error

module Message = Middleware__.Middleware_message

module Gzip = Middleware__.Middleware_gzip

module Security = Middleware__.Middleware_security

module Static = Middleware__.Middleware_static

module Session = Middleware__.Middleware_session

module Authn = Middleware__.Middleware_authn

val error : unit -> t

val static : local_path_f:(unit -> string) -> uri_prefix_f:(unit -> string) -> ?⁠headers:Cohttp.Header.t -> ?⁠etag_of_fname:(string -> string) -> unit -> t

val cookie : unit -> t

val create : name:string -> (Http.Route.handler -> Http.Route.handler) -> t

create ~name (h -> h) create a middleware with name.