Module Queue_core.WorkableJob

A workable job can process a job instance that is persisted. We can not store the job directly because of the polymorphic type ('a Job.t).

type t = {

name : Base.string;

with_context : Core.Ctx.t -> Core.Ctx.t;

work : Core.Ctx.t -> input:Base.string Base.option -> (Base.unit, Base.string) Base.Result.t Lwt.t;

failed : Core.Ctx.t -> (Base.unit, Base.string) Base.Result.t Lwt.t;

max_tries : Base.int;

retry_delay : Utils.Time.duration;

}

val retry_delay : t -> Utils.Time.duration

val max_tries : t -> Base.int

val failed : t -> Core.Ctx.t -> (Base.unit, Base.string) Base.Result.t Lwt.t

val work : t -> Core.Ctx.t -> input:Base.string Base.option -> (Base.unit, Base.string) Base.Result.t Lwt.t

val with_context : t -> Core.Ctx.t -> Core.Ctx.t

val name : t -> Base.string

module Fields : sig ... end

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val of_job : 'a Job.t -> t