Module Queue__Model

module Job : sig ... end

This is the description of a job. A job dispatch is a job description and some arguments/input.

module WorkableJob : sig ... end

A workable job can process a job instance that is persisted. We can not store the job directly because of the polymorphic type ('a Job.t).

module JobInstance : sig ... end

This is the actual job instance that is derived from the job description 'a Job.t and some input. This needs to be serialized and persisted for persistent job queues.