Module Queue__Queue_service.MakePolling

Parameters

Log : Log.Service.Sig.SERVICE

ScheduleService : Schedule.Service.Sig.SERVICE

Repo : Sig.REPO

Signature

include Core.Container.SERVICE

val lifecycle : Core.Container.Lifecycle.t

val dispatch : Core.Ctx.t -> job:'a Sig.Job.t -> ?⁠delay:Utils.Time.duration -> 'a -> unit Lwt.t

Queue a job for processing. Use delay to run the initially job after a certain amount of time.

val register_jobs : Core.Ctx.t -> jobs:'a Sig.Job.t list -> unit Lwt.t

Register jobs that can be dispatched.

Only registered jobs can be dispatched. Dispatching a job that was not registered does nothing.