Module Server_service.MakeOpium

Parameters

Log : Log.Service.Sig.SERVICE

CmdService : Cmd.Service.Sig.SERVICE

Signature

include Core.Container.SERVICE

val lifecycle : Core.Container.Lifecycle.t

val register_endpoints : Server__.Server_core.endpoint list -> unit

Register HTTP routes that are served by the web server.

val start_server : Core.Ctx.t -> unit Lwt.t

Start a HTTP server.

This Lwt.t resolves immediately and the web server starts in the background. The web server serves the registered routes.