Module Sihl.Container

include module type of Sihl__.Core_container

Lifecycle

Every service has a lifecycle, meaning it can be started and stopped. *

type lifecycle = {

name : string;

dependencies : unit -> lifecycle list;

start : unit -> unit Lwt.t;

stop : unit -> unit Lwt.t;

}

val create_lifecycle : ?⁠dependencies:(unit -> lifecycle list) -> ?⁠start:(unit -> unit Lwt.t) -> ?⁠stop:(unit -> unit Lwt.t) -> string -> lifecycle

Service

A service has a start and stop functions and a lifecycle. *

module Service = Sihl__.Core_container.Service

val start_services : Service.t list -> lifecycle list Lwt.t

start_services lifecycles starts a list of service lifecycles. The order does not matter as the services are started in the order of their dependencies. (No service is started before its dependency)

val stop_services : Service.t list -> unit Lwt.t

stop_services ctx services stops a list of service lifecycles with a context ctx. The order does not matter as the services are stopped in the order of their dependencies. (No service is stopped after its dependency)

module Map = Sihl__.Core_container.Map

val collect_all_lifecycles : lifecycle list -> lifecycle Map.t

val top_sort_lifecycles : lifecycle list -> lifecycle list

val unpack : string -> ?⁠default:'a -> 'a option Stdlib.ref -> 'a