Module Sihl.Email

include Sihl_facade.Email

val to_sexp : Sihl_contract.Email.t -> Sexplib0.Sexp.t

val to_yojson : Sihl_contract.Email.t -> Yojson.Safe.t

val of_yojson : Yojson.Safe.t -> Sihl_contract.Email.t option

val pp : Stdlib.Format.formatter -> Sihl_contract.Email.t -> unit

val set_text : string -> Sihl_contract.Email.t -> Sihl_contract.Email.t

val set_html : string option -> Sihl_contract.Email.t -> Sihl_contract.Email.t

val create : ?⁠html:string -> ?⁠cc:string list -> ?⁠bcc:string list -> sender:string -> recipient:string -> subject:string -> string -> Sihl_contract.Email.t

include Sihl_contract.Email.Sig

include Sihl_core.Container.Service.Sig

val lifecycle : Sihl_core.Container.Lifecycle.t

val inbox : unit -> Sihl_contract.Email.t list

inbox () returns the content of the development in-memory mailbox. Intercepted emails land here, they can be used during testing to make sure that certain emails were sent.

val clear_inbox : unit -> unit

clear_inbox () removes all the emails from the development in-memory mailbox. A subsequent call on `inbox ()` will return an empty list.

val send : Sihl_contract.Email.t -> unit Lwt.t

send email sends the email email. The returned Lwt.t fulfills if the underlying email transport acknowledges sending. In case of SMTP, this might take a while.

val bulk_send : Sihl_contract.Email.t list -> unit Lwt.t

bulk_send emails Sends the list of emails emails. If sending of one of them fails, the returned Lwt.t fails.

val register : unit -> Sihl_core.Container.Service.t

val lifecycle : unit -> Sihl_core.Container.Lifecycle.t

val register : (module Sihl_contract.Email.Sig) -> Sihl_core.Container.Service.t

module Setup : sig ... end