Module Sihl.Log

Use Log to set up a logger for your Sihl app. This module can not be used to actually log, use Logs for that.

include module type of Sihl__.Core_log

val get_log_level : unit -> Logs.level option

val logs_dir : unit -> string

val lwt_file_reporter : unit -> Logs.reporter

val format_reporter : ?⁠pp_header:(Logs.src -> Stdlib.Format.formatter -> (Logs.level * string option) -> unit) -> ?⁠app:Stdlib.Format.formatter -> ?⁠dst:Stdlib.Format.formatter -> unit -> Logs.reporter

val cli_reporter : ?⁠pp_header:(Logs.src -> Stdlib.Format.formatter -> (Logs.level * string option) -> unit) -> ?⁠app:Stdlib.Format.formatter -> ?⁠dst:Stdlib.Format.formatter -> unit -> Logs.reporter

val default_reporter : Logs.reporter