Module User.Password_reset

include Sihl_facade.Password_reset

include Sihl_contract.Password_reset.Sig

include Sihl_core.Container.Service.Sig

val lifecycle : Sihl_core.Container.Lifecycle.t

val create_reset_token : email:string -> string option Lwt.t

Create and store a reset token.

Returns None if there is no user with email. The reset token can be used with reset_password to set the password without knowing the old password.

val reset_password : token:string -> password:string -> password_confirmation:string -> (unit, string) Result.t Lwt.t

Set the password of a user associated with the reset token.

val register : unit -> Sihl_core.Container.Service.t

val lifecycle : unit -> Sihl_core.Container.Lifecycle.t

val register : (module Sihl_contract.Password_reset.Sig) -> Sihl_core.Container.Service.t

module Setup : sig ... end