Module Web.Flash

val find_alert : Rock.Request.t -> string option

find_alert request returns the alert stored in the flash storage of the current request.

Make sure that the flash middleware is installed.

val set_alert : string -> Rock.Response.t -> Rock.Response.t

set_alert alert response returns a response with an alert message associated to it. Use alert to tell the user that something went wrong.

Make sure that the flash middleware is installed.

val find_notice : Rock.Request.t -> string option

find_notice request returns the notice stored in the flash storage of the current request.

Make sure that the flash middleware is installed.

val set_notice : string -> Rock.Response.t -> Rock.Response.t

set_notice notice response returns a response with a notice message associated to it. Use notice to tell the user that something happened (successfully).

Make sure that the flash middleware is installed.

val find : string -> Rock.Request.t -> string option

find key request returns the string stored in the flash storage of the current request associated with key.

Make sure that the flash middleware is installed.

val set : (string * string) list -> Rock.Response.t -> Rock.Response.t

set flash response returns a response with the flash stored. Use this to store arbitrary key-value values in the flash store.

Make sure that the flash middleware is installed.