Module Web.Request

include module type of Opium.Request

type t = Rock.Request.t = {

version : Opium__.Version.t;

target : string;

headers : Opium__.Headers.t;

meth : Opium__.Method.t;

body : Opium__.Body.t;

env : Opium__.Context.t;

}

val make : ?⁠version:Opium__.Version.t -> ?⁠body:Opium__.Body.t -> ?⁠env:Opium__.Context.t -> ?⁠headers:Opium__.Headers.t -> string -> Opium__.Method.t -> t

val get : ?⁠version:Opium__.Version.t -> ?⁠body:Opium__.Body.t -> ?⁠env:Opium__.Context.t -> ?⁠headers:Opium__.Headers.t -> string -> t

val post : ?⁠version:Opium__.Version.t -> ?⁠body:Opium__.Body.t -> ?⁠env:Opium__.Context.t -> ?⁠headers:Opium__.Headers.t -> string -> t

val put : ?⁠version:Opium__.Version.t -> ?⁠body:Opium__.Body.t -> ?⁠env:Opium__.Context.t -> ?⁠headers:Opium__.Headers.t -> string -> t

val delete : ?⁠version:Opium__.Version.t -> ?⁠body:Opium__.Body.t -> ?⁠env:Opium__.Context.t -> ?⁠headers:Opium__.Headers.t -> string -> t

val of_plain_text : ?⁠version:Opium__.Version.t -> ?⁠headers:Opium__.Headers.t -> ?⁠env:Opium__.Context.t -> body:string -> string -> Opium__.Method.t -> t

val of_json : ?⁠version:Opium__.Version.t -> ?⁠headers:Opium__.Headers.t -> ?⁠env:Opium__.Context.t -> body:Yojson.Safe.t -> string -> Opium__.Method.t -> t

val of_urlencoded : ?⁠version:Opium__.Version.t -> ?⁠headers:Opium__.Headers.t -> ?⁠env:Opium__.Context.t -> body:(string * string list) list -> string -> Opium__.Method.t -> t

val to_plain_text : t -> string Lwt.t

val to_json : t -> Yojson.Safe.t option Lwt.t

val to_json_exn : t -> Yojson.Safe.t Lwt.t

val to_urlencoded : t -> (string * string list) list Lwt.t

val to_multipart_form_data : ?⁠callback:(name:string -> filename:string -> string -> unit Lwt.t) -> t -> (string * string) list option Lwt.t

val to_multipart_form_data_exn : ?⁠callback:(name:string -> filename:string -> string -> unit Lwt.t) -> t -> (string * string) list Lwt.t

val header : string -> t -> string option

val headers : string -> t -> string list

val add_header : (string * string) -> t -> t

val add_header_or_replace : (string * string) -> t -> t

val add_header_unless_exists : (string * string) -> t -> t

val add_headers : (string * string) list -> t -> t

val add_headers_or_replace : (string * string) list -> t -> t

val add_headers_unless_exists : (string * string) list -> t -> t

val remove_header : string -> t -> t

val content_type : t -> string option

val set_content_type : string -> t -> t

val authorization : t -> Opium__.Auth.Credential.t option

val set_authorization : Opium__.Auth.Credential.t -> t -> t

val cookie : ?⁠signed_with:Opium__.Cookie.Signer.t -> string -> t -> string option

val cookies : ?⁠signed_with:Opium__.Cookie.Signer.t -> t -> Opium__.Cookie.value list

val add_cookie : ?⁠sign_with:Opium__.Cookie.Signer.t -> Opium__.Cookie.value -> t -> t

val add_cookie_unless_exists : ?⁠sign_with:Opium__.Cookie.Signer.t -> Opium__.Cookie.value -> t -> t

val remove_cookie : string -> t -> t

val urlencoded : string -> t -> string option Lwt.t

val urlencoded_exn : string -> t -> string Lwt.t

val urlencoded_list : string -> t -> string list Lwt.t

val query : string -> t -> string option

val query_exn : string -> t -> string

val query_list : t -> (string * string list) list

val sexp_of_t : t -> Sexplib0.Sexp.t

val pp : Stdlib.Format.formatter -> t -> unit

val pp_hum : Stdlib.Format.formatter -> t -> unit

val bearer_token : t -> string option

bearer_token request returns the Bearer token in the Authorization header. The header value has to be of form Bearer <token>.