Module KERNEL.Db

include Sihl__.Core_container.SERVICE

val lifecycle : Sihl__.Core_container.Lifecycle.t

val create_pool : unit -> Sihl__.Data_db_core.pool

Create a database connection pool. Raises Data_db_core.Exception.

val ctx_with_pool : unit -> Sihl__.Core_ctx.t

Create a database connection pool and attach the pool to an empty context.

val add_pool : Sihl__.Core_ctx.t -> Sihl__.Core_ctx.t

Create a database connection pool and attach to provided context.

val query : Sihl__.Core_ctx.t -> (Caqti_lwt.connection -> ('a, Caqti_error.t) Result.t Lwt.t) -> 'a Lwt.t

Run a database query.

The context has to contain a database connection or a database connection pool.

val with_connection : Sihl__.Core_ctx.t -> (Sihl__.Core_ctx.t -> 'a Lwt.t) -> 'a Lwt.t

Run a database query on a single connection. Can be used to set session variables that are bound to the same connection.

val atomic : Sihl__.Core_ctx.t -> (Sihl__.Core_ctx.t -> 'a Lwt.t) -> 'a Lwt.t

Run a database query atomically on a connection.

The context has to contain a database connection or a database connection pool. Fetch a database connection from context if necessary to make sure, that every query runs on the same connection.

val with_disabled_fk_check : Sihl.Core.Ctx.t -> (Sihl__.Core_ctx.t -> 'a Lwt.t) -> 'a Lwt.t

Disables foreign key checks if supported by the database.

Use very carefully, data might become inconsistent!