Module Authn_service.Make

Parameters

SessionService : Sihl.Session.Sig.SERVICE

UserService : Sihl.User.Sig.SERVICE

Signature

include Sihl__.Core_container.SERVICE

val lifecycle : Sihl__.Core_container.Lifecycle.t

val find_user_in_session_opt : Sihl.Core.Ctx.t -> Sihl.User.t option Lwt.t

Find currently logged in user in the current context.

Make sure to call authenticate_session before or apply the required session and authentication middlewares.

val find_user_in_session : Sihl.Core.Ctx.t -> Sihl.User.t Lwt.t

Find currently logged in user in the current context.

Make sure to call authenticate_session before or apply the required session and authentication middlewares.

val authenticate_session : Sihl.Core.Ctx.t -> Sihl.User.t -> unit Lwt.t

Assign a user to the current anonymous session.

Use authenticate_session ctx user to log in a user. If a user is already assigned to session, replace the user.

val unauthenticate_session : Sihl.Core.Ctx.t -> unit Lwt.t

Log user out.

Remove user from current session so that session is anonymous again.