Module type Contract_cache.Sig

val set : (string * string option) -> unit Lwt.t

set entry inserts an entry into the cache storage. entry is a tuple where the first element is the key and the second element is the value. Since the value is an optional, set can be used to remove a value from the store like so: set ("foo", None). If a key exists already, the value is overwritten with the provided value.

val find : string -> string option Lwt.t

find key returns the value that is associated with key.

val register : unit -> Sihl__.Core_container.Service.t

include Sihl__.Core_container.Service.Sig

val lifecycle : Sihl__.Core_container.lifecycle