Module Database_migration.Make

Parameters

Repo : Sihl__.Database_migration_repo.Sig

Signature

val register_migration : Sihl__.Contract_migration.t -> unit

register_migration migration registers a migration migration with the migration service so it can be executed with `run_all`.

val register_migrations : Sihl__.Contract_migration.t list -> unit

register_migrations migrations registers migrations migrations with the migration service so it can be executed with `run_all`.

val execute : Sihl__.Contract_migration.t list -> unit Lwt.t

execute migrations runs all migrations migrations.

val run_all : unit -> unit Lwt.t

run_all () runs all migrations that have been registered.

val pending_migrations : unit -> (string * int) list Lwt.t

pending_migrations () returns a list of migrations that need to be executed in order to have all migrations applied. The returned migration is a tuple (namespace, number) where namespace is the namespace of the migration and number is the number of pending migrations that need to be applied in order to achieve the desired schema version.

An empty list means that there are no pending migrations and that the database schema is up-to-date.

val register : Sihl__.Contract_migration.t list -> Sihl__.Core_container.Service.t

include Sihl__.Core_container.Service.Sig

val lifecycle : Sihl__.Core_lifecycle.lifecycle