Module Sihl__.Email_service

val let* : ('a, 'b) Lwt_result.t -> ('a -> ('c, 'b) Lwt_result.t) -> ('c, 'b) Lwt_result.t

module EnvConfigProvider : Sihl__.Email_sig.ConfigProvider.SMTP

module Template : sig ... end

module Make : sig ... end

module MakeDelayed : functor (EmailService : Sihl__.Email_sig.SERVICE) -> functor (DbService : Sihl.Data.Db.Sig.SERVICE) -> functor (QueueService : Sihl__.Queue_sig.SERVICE) -> Sihl__.Email_sig.SERVICE

Use this functor to create an email service that sends emails using the job queue. This is useful if you need to answer a request quickly while sending the email in the background