Module type Sihl__Contract_password_reset.Sig

val create_reset_token : email:string -> string option Lwt.t

Create and store a reset token.

Returns None if there is no user with email. The reset token can be used with reset_password to set the password without knowing the old password.

val reset_password : token:string -> password:string -> password_confirmation:string -> (unit, string) Result.t Lwt.t

Set the password of a user associated with the reset token.

val register : unit -> Sihl__.Core_container.Service.t

include Sihl__.Core_container.Service.Sig

val lifecycle : Sihl__.Core_lifecycle.lifecycle