Module type Sihl__Contract_queue.Sig

val dispatch : 'a t -> ?⁠delay:Sihl__.Core_time.duration -> 'a -> unit Lwt.t

dispatch job ?delay input queues job for processing while input input is the input that the job needs to run. Use delay to run the job earliest after a certain amount of time.

val register_jobs : 'a t list -> unit Lwt.t

register_jobs jobs registers jobs that can be dispatched.

Only registered jobs can be dispatched. Dispatching a job that was not registered does nothing.

val register : ?⁠jobs:'a t list -> unit -> Sihl__.Core_container.Service.t

include Sihl__.Core_container.Service.Sig

val lifecycle : Sihl__.Core_container.lifecycle