Module type Sihl__Contract_random.Sig

val base64 : int -> string

base64 n returns a Base64 encoded string containing n random bytes.

val bytes : int -> string

bytes n returns a byte sequence as string with n random bytes. In most cases you want to use base64 to get a string that can be used safely in most web contexts.

val register : unit -> Sihl__.Core_container.Service.t

include Sihl__.Core_container.Service.Sig

val lifecycle : Sihl__.Core_lifecycle.lifecycle