Module Sihl__Core_random

val base64 : int -> string

base64 n returns a Base64 encoded string containing n random bytes.

val bytes : int -> char list

bytes n returns a byte sequence as string with n random bytes. In most cases you want to use base64 to get a string that can be used safely in most web contexts.