Module Sihl__Email_service

module EnvConfigProvider : functor (Config : Sihl__.Config_sig.SERVICE) -> Sihl__.Email_sig.ConfigProvider.SMTP

module Template : sig ... end

module Make : sig ... end

module MakeDelayed : functor (Log : Sihl.Log.Sig.SERVICE) -> functor (EmailService : Sihl__.Email_sig.SERVICE) -> functor (DbService : Sihl.Data.Db.Sig.SERVICE) -> functor (QueueService : Sihl__.Queue_sig.SERVICE) -> Sihl__.Email_sig.SERVICE

Use this functor to create an email service that sends emails using the job queue. This is useful if you need to answer a request quickly while sending the email in the background