Module Sihl__Email_service.MakeDelayed

Use this functor to create an email service that sends emails using the job queue. This is useful if you need to answer a request quickly while sending the email in the background

Parameters

EmailService : Sihl__.Email_sig.SERVICE

DbService : Sihl.Data.Db.Sig.SERVICE

QueueService : Sihl__.Queue_sig.SERVICE

Signature

include Sihl__.Core_container.SERVICE

val lifecycle : Sihl__.Core_container.Lifecycle.t

module Template : Sihl__.Email_sig.Template.SERVICE

val send : Sihl.Core.Ctx.t -> Sihl__.Email_core.t -> unit Lwt.t

val bulk_send : Sihl.Core.Ctx.t -> Sihl__.Email_core.t list -> unit Lwt.t