Module type ConfigProvider.SMTP

Read the configurations from environment variables and set sane defaults. SMTP_SENDER: Sender address from where the emails come from SMTP_HOST: Host address of the SMTP server SMTP_USERNAME: Username for the SMTP server login SMTP_USERNAME: Password for the SMTP server login SMTP_PORT: Port number, default is 587 SMTP_START_TLS: Whether to use TLS, default is true CA_DIR: Location of root CA certificates on the file system, default is /etc/ssl/certs

val sender : Sihl.Core.Ctx.t -> (string, string) Lwt_result.t

val username : Sihl.Core.Ctx.t -> (string, string) Lwt_result.t

val password : Sihl.Core.Ctx.t -> (string, string) Lwt_result.t

val host : Sihl.Core.Ctx.t -> (string, string) Lwt_result.t

val port : Sihl.Core.Ctx.t -> (int option, string) Lwt_result.t

val start_tls : Sihl.Core.Ctx.t -> (bool, string) Lwt_result.t

val ca_dir : Sihl.Core.Ctx.t -> (string, string) Lwt_result.t