Module Sihl__User_password_reset_service.Make

Parameters

TokenService : Sihl.Token.Sig.SERVICE

UserService : Sihl__.User_sig.SERVICE

Signature

include Sihl.Core.Container.SERVICE

val lifecycle : Sihl.Core.Container.Lifecycle.t

val create_reset_token : Sihl.Core.Ctx.t -> email:string -> Sihl.Token.t option Lwt.t

Create and store a reset token.

Returns None if there is no user with email. The reset token can be used with reset_password to set the password without knowing the old password.

val reset_password : Sihl.Core.Ctx.t -> token:string -> password:string -> password_confirmation:string -> (unit, string) Result.t Lwt.t

Set the password of a user associated with the reset token.