Module type Sihl__Web_server_sig.SERVICE

include Sihl__.Core_container.SERVICE

val lifecycle : Sihl__.Core_container.Lifecycle.t

val register_endpoints : Sihl__.Web_server_core.endpoint list -> unit

Register HTTP routes that are served by the web server.

val start_server : Sihl.Core.Ctx.t -> unit Lwt.t

Start a HTTP server.

This Lwt.t resolves immediately and the web server starts in the background. The web server serves the registered routes.