Module Sihl_authn__Service.Make

Parameters

SessionService : Session.Sig.SERVICE

UserService : User.Sig.SERVICE

Signature

include Sihl_authn.Sig.Core.Container.Service.Sig

val lifecycle : Sihl_core.Container.Lifecycle.t

val find_user_in_session_opt : Sihl_authn.Sig.Core.Ctx.t -> Sihl_authn.Sig.Session.t -> Sihl_authn.Sig.User.t option Lwt.t

Find currently logged in user in the current context.

Make sure to call authenticate_session before or apply the required session and authentication middlewares.

val find_user_in_session : Sihl_authn.Sig.Core.Ctx.t -> Sihl_authn.Sig.Session.t -> Sihl_authn.Sig.User.t Lwt.t

Find currently logged in user in the current context.

Make sure to call authenticate_session before or apply the required session and authentication middlewares.

val authenticate_session : Sihl_authn.Sig.Core.Ctx.t -> Sihl_authn.Sig.User.t -> Sihl_authn.Sig.Session.t -> unit Lwt.t

Assign a user to the current anonymous session.

Use authenticate_session ctx user to log in a user. If a user is already assigned to session, replace the user.

val unauthenticate_session : Sihl_authn.Sig.Core.Ctx.t -> Sihl_authn.Sig.Session.t -> unit Lwt.t

Log user out.

Remove user from current session so that session is anonymous again.

val register : unit -> Sihl_authn.Sig.Core.Container.Service.t