Module Sihl_core

This is the core of Sihl, every other component builds on top of it.

A module to manage CLI commands, configurations, service lifecycles, the request context, and Sihl apps.

module Container : sig ... end

A module to manage the service container and service lifecycles.

module App : sig ... end

A module to create Sihl apps.

module Configuration : sig ... end

A module to manage service configurations.

module Command : sig ... end

A module to register and run CLI commands.

module Log : sig ... end

module Random : sig ... end