Module Sihl_database.Service

include Sig.Core.Container.Service.Sig

val lifecycle : Sihl_core.Container.Lifecycle.t

val raise_error : ('a, Caqti_error.t) Result.t -> 'a

raise_error err raises a printable caqti error err .

val fetch_pool : unit -> Sihl_database__.Model.pool

fetch_pool () returns the connection pool.

val query : (Caqti_lwt.connection -> 'a Lwt.t) -> 'a Lwt.t

query ctx f runs the query f on the connection pool and returns the result. If the query fails the Lwt.t fails as well.

val transaction : (Caqti_lwt.connection -> 'a Lwt.t) -> 'a Lwt.t

transaction ctx f runs the query f on the connection pool in a transaction and returns the result. If the query fails the Lwt.t fails as well and the transaction gets rolled back. If the database driver doesn't support transactions, transaction gracefully becomes query.

val register : unit -> Sig.Core.Container.Service.t