Module Service.Make

Parameters

SessionService : Session.Sig.SERVICE

Signature

val fetch_entry : Sig.Session.t -> Sihl_message__.Model.Entry.t option Lwt.t

val find_current : Sig.Session.t -> Sihl_message__.Model.Message.t option Lwt.t

val set_next : Sig.Session.t -> Sihl_message__.Model.Message.t -> unit Lwt.t

val rotate : Sig.Session.t -> Sihl_message__.Model.Message.t option Lwt.t

val current : Sig.Session.t -> Sihl_message__.Model.Message.t option Lwt.t

val set : Sig.Session.t -> ?⁠error:string list -> ?⁠warning:string list -> ?⁠success:string list -> ?⁠info:string list -> unit -> unit Lwt.t

Set flash message for the current session.

Flash messages can be used to transport information across request response lifecycles. The typical use case is giving a user feedback after a form submission.

Requires middlewares: Session & Message

val register : unit -> Sig.Core.Container.Service.t