Module Sihl_middleware

A middleware is a function that takes a Http.Route.handler and returns a Http.Route.handler. It is typically used to add content to the request context that is valid only during a request.

module Clickjacking = Sihl_middleware__.Middleware_clickjacking

module Cookie = Sihl_middleware__.Middleware_cookie

module Csrf = Sihl_middleware__.Middleware_csrf

module Error = Sihl_middleware__.Middleware_error

module Message = Sihl_middleware__.Middleware_message

module Gzip = Sihl_middleware__.Middleware_gzip

module Security = Sihl_middleware__.Middleware_security

module Static = Sihl_middleware__.Middleware_static

module Session = Sihl_middleware__.Middleware_session

module Authn = Sihl_middleware__.Middleware_authn

val error : unit -> Opium_kernel.Rock.Middleware.t

val static : local_path_f:(unit -> string) -> uri_prefix_f:(unit -> string) -> ?⁠headers:Cohttp.Header.t -> ?⁠etag_of_fname:(string -> string) -> unit -> Opium_kernel.Rock.Middleware.t

val cookie : unit -> Opium_kernel.Rock.Middleware.t