Parameter MakePostgreSql.1-Database

include Database.Sig.Core.Container.Service.Sig

val lifecycle : Sihl_core.Container.Lifecycle.t

val fetch_pool : unit -> Sihl_database__.Model.pool

fetch_pool () returns the connection pool.

val query : Database.Sig.Core.Ctx.t -> (Caqti_lwt.connection -> 'a Lwt.t) -> 'a Lwt.t

query ctx f runs the query f on the connection pool and returns the result. If the query fails the Lwt.t fails as well.

val transaction : Database.Sig.Core.Ctx.t -> (Caqti_lwt.connection -> 'a Lwt.t) -> 'a Lwt.t

transaction ctx f runs the query f on the connection pool in a transaction and returns the result. If the query fails the Lwt.t fails as well and the transaction gets rolled back. If the database driver doesn't support transactions, transaction gracefully becomes query.

val register : unit -> Database.Sig.Core.Container.Service.t