Module Service.Make

Parameters

TokenService : Token.Sig.SERVICE

UserService : User.Sig.SERVICE

Signature

include Sig.Core.Container.Service.Sig

val lifecycle : Sihl_core.Container.Lifecycle.t

val create_reset_token : Sig.Core.Ctx.t -> email:string -> Sig.Token.t option Lwt.t

Create and store a reset token.

Returns None if there is no user with email. The reset token can be used with reset_password to set the password without knowing the old password.

val reset_password : Sig.Core.Ctx.t -> token:string -> password:string -> password_confirmation:string -> (unit, string) Result.t Lwt.t

Set the password of a user associated with the reset token.

val configure : Sig.Core.Configuration.data -> Sig.Core.Container.Service.t