Module type Sig.SERVICE

include Core.Container.Service.Sig

val lifecycle : Sihl_core.Container.Lifecycle.t

val dispatch : job:'a Job.t -> ?⁠delay:Utils.Time.duration -> 'a -> unit Lwt.t

Queue a job for processing. Use delay to run the initially job after a certain amount of time.

val register_jobs : jobs:'a Job.t list -> unit Lwt.t

Register jobs that can be dispatched.

Only registered jobs can be dispatched. Dispatching a job that was not registered does nothing.

val register : ?⁠jobs:'a Job.t list -> unit -> Core.Container.Service.t