Module Sihl_queue_core__Model

module Utils = Sihl_utils

module Database = Sihl_database

module Job : sig ... end

module WorkableJob : sig ... end

A workable job can process a job instance that is persisted. We can not store the job directly because of the polymorphic type ('a Job.t).

module JobInstance : sig ... end

This is the actual job instance that is derived from the job description 'a Job.t and some input. This needs to be serialized and persisted for persistent job queues.