Module Sihl_schedule

Schedules are functions that need to run periodically, similarly to cron jobs. Use this for instance to run clean up jobs.

module Service : Sig.SERVICE

module Sig : sig ... end

type t = Sihl_schedule__.Model.t

val create : Sihl_schedule__.Model.scheduled_time -> f:(unit -> unit Lwt.t) -> label:string -> t

val every_second : Sihl_schedule__.Model.scheduled_time

val every_hour : Sihl_schedule__.Model.scheduled_time