Module Service.Make

Parameters

Repo : Sihl_token.Sig.REPOSITORY

Signature

include Sihl_token.Sig.Core.Container.Service.Sig

val lifecycle : Sihl_core.Container.Lifecycle.t

val create : kind:string -> ?⁠data:string -> ?⁠expires_in:Sihl_token.Sig.Utils.Time.duration -> ?⁠length:int -> unit -> Sihl_token__.Model.t Lwt.t

Create a token and store a token.

Provide expires_in to define a duration in which the token is valid, default is one day. Provide data to store optional data as string. Provide length to define the length of the token in bytes.

val find : string -> Sihl_token__.Model.t Lwt.t

Returns an active and non-expired token. Raises Failure if no token is found.

val find_opt : string -> Sihl_token__.Model.t option Lwt.t

Returns an active and non-expired token.

val find_by_id : string -> Sihl_token__.Model.t Lwt.t

Returns an active and non-expired token by id. Raises Failure if no token is found.

val find_by_id_opt : string -> Sihl_token__.Model.t option Lwt.t

Returns an active and non-expired token by id.

val invalidate : Sihl_token__.Model.t -> unit Lwt.t

Invalidate a token by marking it as such in the database and therefore marking it "to be deleted"

val register : unit -> Sihl_token.Sig.Core.Container.Service.t