Module Storage

Use this module to store and retrieve large files. This is typically used for binary files such as images or audio.

module Service = Storage__.Storage_service

module File = Storage__.Storage_core.File

module StoredFile = Storage__.Storage_core.StoredFile