Module User_service.Make

Parameters

Log : Log.Service.Sig.SERVICE

CmdService : Cmd.Service.Sig.SERVICE

DbService : Data.Db.Service.Sig.SERVICE

Repo : Sig.REPOSITORY

Signature

include Core.Container.SERVICE

val lifecycle : Core.Container.Lifecycle.t

val start : Core.Container.start

val stop : Core.Container.stop

val add_user : User__.User_core.User.t -> Core.Ctx.t -> Core.Ctx.t

val require_user_opt : Core.Ctx.t -> User__.User_core.User.t option

val require_user : Core.Ctx.t -> User__.User_core.User.t

val find_all : Core.Ctx.t -> query:Data.Ql.t -> (User__.User_core.User.t list * Data.Repo.Meta.t) Lwt.t

val find_opt : Core.Ctx.t -> user_id:string -> User__.User_core.User.t option Lwt.t

val find : Core.Ctx.t -> user_id:string -> User__.User_core.User.t Lwt.t

val find_by_email : Core.Ctx.t -> email:string -> User__.User_core.User.t Lwt.t

val find_by_email_opt : Core.Ctx.t -> email:string -> User__.User_core.User.t option Lwt.t

val update_password : Core.Ctx.t -> ?⁠password_policy:(string -> (unit, string) Result.t) -> user:User__.User_core.User.t -> old_password:string -> new_password:string -> new_password_confirmation:string -> unit -> (User__.User_core.User.t, string) Result.t Lwt.t

val update_details : Core.Ctx.t -> user:User__.User_core.User.t -> email:string -> username:string option -> User__.User_core.User.t Lwt.t

val set_password : Core.Ctx.t -> ?⁠password_policy:(string -> (unit, string) Result.t) -> user:User__.User_core.User.t -> password:string -> password_confirmation:string -> unit -> (User__.User_core.User.t, string) Result.t Lwt.t

Set the password of a user without knowing the old password.

This feature is typically used by admins.

val create_user : Core.Ctx.t -> email:string -> password:string -> username:string option -> User__.User_core.User.t Lwt.t

Create and store a user.

val create_admin : Core.Ctx.t -> email:string -> password:string -> username:string option -> User__.User_core.User.t Lwt.t

Create and store a user that is also an admin.

val register : Core.Ctx.t -> ?⁠password_policy:(string -> (unit, string) Stdlib.result) -> ?⁠username:string -> email:string -> password:string -> password_confirmation:string -> unit -> (User__.User_core.User.t, string) Result.t Lwt.t

Create and store new user.

Provide password_policy to check whether the password fulfills certain criteria.

val login : Core.Ctx.t -> email:string -> password:string -> (User__.User_core.User.t, string) Result.t Lwt.t

Find user by email if password matches.