Sihl Email

This module provides the email service and the email template service.

Email service

The email service provides API for sending emails. Currently SMTP and Sendgrid are supported.

Installation

Backend

First, decide for a backend in service/service.ml:

module Email = Sihl_email.SMTP

Registration

Then, register the service in run/run.ml:

let services = [ Service.Email.register () ]

Configuration

Run make sihl to get a list of required configurations.

Usage

The API is documented in Sihl.Contract.Email.Sig.

Template service

The template service can be used to create and edit email templates. Email templates are rendered with parameters to emails, that can be sent with the email service. Currently MariaDb (Sihl_cache.MariaDb) and PostgreSql (Sihl_cache.PostgreSql) are supported.

Installation

Backend

First, decide for a backend in service/service.ml:

module Email_template = Sihl_email.Temlate.PostgreSql

Registration

Then, register the service in run/run.ml:

let services = [ Service.Email_template.register () ]

Migrations

Run make sihl migrate to run pending migrations.

Usage

The API is documented in Sihl.Contract.Email_template.Sig.