Module Core

This is the core of Sihl, every other component builds on top of it.

A module to manage CLI commands, configurations, service lifecycles, the request context, and Sihl apps.

module Container : sig ... end

A module to manage the service container and service lifecycles.

module Ctx : sig ... end

The service request context is a key value store that can be used to pass arbitrary values from one service to another within the same request.

module App : sig ... end

A module to create Sihl apps.

module Configuration : sig ... end

A module to manage service configurations.

module Command : sig ... end

A module to register and run CLI commands.

module Log : sig ... end