Module Middleware

A middleware is a function that takes a Http.Route.handler and returns a Http.Route.handler. It is typically used to add content to the request context that is valid only during a request.

module Clickjacking = Middleware__.Middleware_clickjacking

module Cookie = Middleware__.Middleware_cookie

module Csrf = Middleware__.Middleware_csrf

module Error = Middleware__.Middleware_error

module Message = Middleware__.Middleware_message

module Gzip = Middleware__.Middleware_gzip

module Security = Middleware__.Middleware_security

module Static = Middleware__.Middleware_static

module Session = Middleware__.Middleware_session

module Authn = Middleware__.Middleware_authn

val error : unit -> Opium_kernel.Rock.Middleware.t

val static : local_path_f:(unit -> string) -> uri_prefix_f:(unit -> string) -> ?⁠headers:Cohttp.Header.t -> ?⁠etag_of_fname:(string -> string) -> unit -> Opium_kernel.Rock.Middleware.t

val cookie : unit -> Opium_kernel.Rock.Middleware.t