Module Sihl

module App : sig ... end

This is the main module. Use App app to create Sihl apps.

module Configuration : sig ... end

Use Configuration to read app configuration from environment variables or .env files. Sihl services require a valid configuration at start-up time, they won't start if the configuration they need is not available.

module Web : sig ... end

Use Web to setup routers, handlers and middlewares to deal with HTTP requests.

module Database : sig ... end

Use Database to handle connection pooling, migrations and to query your database.

module Log : sig ... end

Use Log to set up a logger for your Sihl app. This module can not be used to actually log, use Logs for that.

module Cleaner : sig ... end

Use Cleaner to clean persisted service state. This is useful for cleaning the state before running tests.

module Command : sig ... end

module Container : sig ... end

module Time : sig ... end

module Schedule : sig ... end

module Random : sig ... end

module Contract : sig ... end