Module Storage

Use this module to store and retrieve large files. This is typically used for binary files such as images or audio.

exception Exception of string

module File = Storage__.Model.File

module StoredFile = Storage__.Model.StoredFile

module Sig : sig ... end